Warn about partially compiled artifacts
If torch.compile compilation fails in the middle, vLLM's cache directory
can be left with a "partially compiled artifact". Then, when we try to
compile again, vLLM may cache hit on the partially compiled artifact and
get very confused.

This PR adds a warning about this. Fixing this bug is in the works (it
requires torch._inductor.standalone_compile, which has stronger
guarantees about partial compilation). I think the best we can do right
now is a warning because it's difficult to detect where in the
compilation we are.

Signed-off-by: rzou <zou3519@gmail.com>