[Bugfix] Fix parallel sampling performance regression when guided decoding is enabled
`ParallelSampleSequenceGroup.add_request` has to copy the original `SamplingParams` instance as
many times as the number of requested samples. This is currently done with a `copy.deepcopy` call,
which is not advisable as the the `logits_processors` field could contain arbitrary Python objects
with expensive-to-copy state. This happens to be the case with the current guided decoding logits
processors, scaling linearly with the value of `SamplingParams.n` and introducing a bottleneck in the
hot path.

A similar issue was previous identified, and `SamplingParams.clone` was introduced to workaround this
issue - it attempts to call a `clone` function on each logits processor object, with the assumption that
classes can implement this method to minimize the overhead by performing shallow copies when possible.
However, not all existing logits processors implement this method. Nor does the `ParallelSampleSequenceGroup`
class avail itself of the `SamplingParams.clone` method.

This commit introduces the following changes:
* Modify `ParallelSampleSequenceGroup.add_request` to call `SamplingParams.clone` instead of `copy.deepcopy`.
* Update the logits processors of the `guidance`, `outlines` and `xgrammar` backends to expose a `clone` method
  for the efficient copying of mutable state.

The `lm-format-enforcer` backend was left untouched as the logits processor implementation is external to
vLLM.

Signed-off-by: Madeesh Kannan <shadeMe@users.noreply.github.com>