remove already used memory from peak memory need computation
If the gpu already is used by torch, e.g. in a RLHF setup, the current peakmemory computation wrongly thinks it does not have enough memory.

Assume we have A100 80Gb and 50Gb is already taken. If the vLLM needs an extra 12 GB and is given 16Gb (gpu_memory_utilization=0.2), the peak memory is measured as 50+12 and therefore throws and error as 62 > 16. It thinks it does not have enough memory while it has. The correct measurement is 12 < 16.