[Security] Prevent new imports of (cloud)pickle
This change proposes an additional pre-commit check that bans new
imports of `pickle` or `cloudpickle` unless they are added to an
allowlist in `tools/check_pickle_imports.py`. The reasoning is that
these modules are known to be unsafe when deserializing data from
potentially untrusted parties. It has resulted in multiple CVEs for
vLLM, and numerous in the python ecosystem more broadly.

There are safer alternatives available. Two relevant alternatives
already in use in vLLM are `msgpack` and `pydantic`. To help prevent
accidental future CVEs, we should do what we can to discourage any
further use of the unsafe choices.

This check is not perfect. It is still possible to add `pickle` imports
using `importlib`, for example. Diligence is still required on the part
of code reviewers to watch out for unsafe additions to the code base.

Signed-off-by: Russell Bryant <rbryant@redhat.com>