fix dummy sequence length setting in llama3.2 (#1229)
In llama3.2, the minimum seq length of dummy sequence is set to 1, which
is not expected in cases that models allow multiple images as encoder
input. Without this PR, the terminal will keep reporting warnings like
below,
```
WARNING 05-08 05:24:25 [profiling.py:242] The sequence length used for profiling 
(max_num_batched_tokens / max_num_seqs = 1) is too short to hold the multi-modal embeddings 
in the worst case (4 tokens in total, out of which {'image': 25616} are reserved for multi-modal embeddings). 
This may cause certain multi-modal inputs to fail during inference, even when the input text is short. 
To avoid this, you should increase `max_model_len`, reduce `max_num_seqs`, and/or reduce `mm_counts`.
```
This PR will eliminate the above warnings.

To reproduce, you need to add `--limit-mm-per-prompt {"images": 4}` in
engine args.

Signed-off-by: zhouyu5 <yu.zhou@intel.com>