Create VideoLoaderRegistry that supports custom implementations of VideoLoader backends.
Summary:
Currently, only OpenCV-based VideoLoader backend is supported, which depends on the opencv-python-headless library. This can sometimes cause dependency conflicts in runtimes. In this diff, we created a VideoLoaderRegistry that allows custom implementations of VideoLoader backends to be plugged in, instead of having a hard dependency on opencv headless.

Test Plan:
- Unit test
- Meta-internal implementation of video loader.

Reviewers:

Subscribers:

Tasks:

Tags: