Disable chunked prefill for pooling models in V1
For encoder-only models the attention mask is not causal
so even if we were to accumulate the hidden states in the
same way as the prompt tokens are accumalated for sampling,
the result would be incorrect.

For example, in some pooling models the CLS token, which is
always the first one, is the one whose hidden states become the
embedding tensor. If we do chunked prefill, it's impossible for
this token to attend to all subsequent tokens.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>