[V1] Update zmq socket creation in nixl connector
While browsing this code, I noticed this code wasn't using the
`make_zmq_socket` helper from `vllm.utils`. Update it to use it.

The benefits are:

- this fixes IPv6 support
- Optimize buffer sizes based on system memory to improve throughput,
  though this part doesn't seem too relevant for how the current sockets
  are used. I still think using the helper is a good best practice.

Fixing IPv6 support also required a fix to how the zmq path string is
constructed to wrap IPv6 addresses in square brackets before adding a
port number to the end.

Signed-off-by: Russell Bryant <rbryant@redhat.com>