Delay resolving of current_platform import by moving it inside _cached_get_attn_backend
Summary:
**Changes:**

* In `selector.py`, the import of `current_platform` from `vllm.platforms` is delayed by moving it inside the `_cached_get_attn_backend` function.
* The delay is conditional, only happening if the `DELAY_CURRENT_PLATFORM_IMPORT` environment variable is set to "1".

**Rationale:**
The change aims to delay the resolution of the `current_platform` import. By moving the import inside the `_cached_get_attn_backend` function, it is only loaded when actually needed. This is necessary to fix issues that arise from importing the platform too early. The conditional delay allows for easy testing and disabling of this behavior by default unless the `DELAY_CURRENT_PLATFORM_IMPORT` environment variable is set to "1".

Differential Revision: D74610999