[V1][Spec Decoding] Use model_loader.get_model() to load models
With:

```
$> vllm serve lmsys/vicuna-7b-v1.3 --speculative-config '{"method": "medusa", "model": "abhigoyal/vllm-medusa-vicuna-7b-v1.3", "num_speculative_tokens": 6}'
```

(A misconfiguration: num_spec_tokens > num_heads)

I noticed in V0 we get this fairly obscure error:

  File "/home/markmc/vllm-project/vllm/vllm/model_executor/model_loader/default_loader.py", line 294, in load_model
    raise ValueError(
ValueError: Following weights were not initialized from checkpoint: {'blocks.5.layers.0.weight', 'lm_heads.5.weight'}

but in V1, we silently accept the config.

It turns out the difference is that in V0 we're using
ModelLoader.load_model() but in V1 we're reimplementing
a subset that logic in the Medusa and EAGLE proposers.

If we use the full model loading infrastructure, we can
reduce the complexity of drafters, and automatically
have support for e.g. quantized weights as per #18290.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>