Add support for prefix caching and chunked prefill
If the attention is causal and the pooler is compatible
with chunking, both prefix caching and chunked prefill
will work with pooling.

In this implementation the pooler sees all chunks
so in theory we could have stateful poolers. However
they would support chunked prefill and not prefix cache
because with prefix cache the prefix chunks aren't seen.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>