[CI] Fix race condition with StatelessProcessGroup.barrier
First, this includes one small security fix to the test_shm_broadcast
test.  Don't use the host's default IP address for this test. Use
localhost.  Otherwise we are binding a socket to an IP address that is
externally reachable. If the host does not have a proper firewall
configuration and is exposed to potentially untrusted traffic, this
would open a short window of time where the host is vulnerable to remote
code execution via pickle deserialization while the test is running.

The real fix included here is to resolve a race condition with the
`barrier()` implementation in `StatelessProcessGroup`. The barrier was
used to synchronize processes in this test, but it did not properly
ensure that the underlying `TCPStore` instance in the
`StatelessProcessGroup` would not be destroyed before other processes in
the group were finished using it. The fix here is a more robust
`barrier()` implementation that does 3 key things:

1. Signal arrival into the barrier.

2. Make all processes wait to receive the signal from all other
   processes that they have arrived in the barrier.

3. For rank 0 only, the one that runs the server side of the `TCPStore`
   instance, ensure it is that last process to exit the barrier by
   waiting until it receives a signal from all other processes that they
   have departed.

Prior to this change, the test failure in issue #18492 was very easy for
me to reproduce - reliably in under 10 tries. With this change in place,
I have not seen the failure occur again (after a couple hundred runs so
far).

Closes #18492

Signed-off-by: Russell Bryant <rbryant@redhat.com>