[BugFix] Increase TP execute_model timeout
The timeout for execute_model (forward pass) running in TP workers was set to 40 seconds but there are a number of reports of this being hit. We assume the workers are in a bad state when this happens and the engine goes into a failure state.

It's not clear in which cases the model forward pass takes this long but for now we can increase the timeout to 5 minutes.

Signed-off-by: Nick Hill <nhill@redhat.com>