Respect passed in device overrides in engine args
Summary:
Signed-off-by: Adolfo Victoria adovi@meta.com

When calling `create_engine_config` we don't respect the device set in engine args and instead grab it from the runtime platform directly. This logic is also extraneous since the default behavior (i.e. device=auto) is to set the device to the runtime platform device if auto is passed inside of `__post_init__` inside of `DeviceConfig`

Differential Revision: D75496412