Improve "failed to get the hash of the compiled graph" error
The most common form of the error is that a previous compilation failed,
leading to a corrupted cache artifact. This error usually hides some
other error (that caused the compilation to fail in the first place), so
we update the error message so that users will loook for the other error
instead.

See #18851 for an example where people are confused and where this error
message would have helped.

We will not automatically delete the cache directory yet; that becomes
unnecessary in PyTorch 2.8.0+ where we switch over to
standalone_compile.

Signed-off-by: rzou <zou3519@gmail.com>