Fix CompilationConfig repr
CompilationConfig repr initially works but when model
compilation runs, it gets dynamically updated with a PostGradPassManager
object:
https://github.com/vllm-project/vllm/blob/main/vllm/compilation/backends.py#L368-L384
This object is not json serializable and causes repr to error out.

This PR excludes the PostGradPassManager object from the repr.

This PR fixes one of the problems mentioned in https://github.com/vllm-project/vllm/issues/18916

Test Plan:
- new test (`pytest tests/test_config.py -v -k "compile_config_repr_succeeds"`)

Signed-off-by: rzou <zou3519@gmail.com>