Remove duplicate kv_b_proj from models using MLA (#1349)
Deepseek in our definition has two places where kv_b_proj is defined: in
`self_attn.kv_b_proj` and `self_attn.impl.kv_b_proj` . First one isn't
used, but at the model initialization is present, which makes inc try to
quantize it. Because at the measurement it wasn't used, there are no
measurements for this specific object and it causes it to crash.

---------

Signed-off-by: kwisniewski98 <kwisniewski@habana.ai>