Export NaNs in logits to scheduler_stats if output is corrupted (#18777)
Summary:
Pull Request resolved: https://github.com/vllm-project/vllm/pull/18777

Signed-off-by: Vlad Mihailescu <vladmihailescu@meta.com>

Report nan in logits in scheduler_stats. This can be used later to bump  Phrometeus counter but for now this is required so we can export it in our internal counter infra.

This counter is used to identify bad hosts or bad GPUs which cause NaNs in logits during model forward passes.

 It's a common metric we expose internally.

Reviewed By: Adolfo-Karim

Differential Revision: D75423285

Signed-off-by: Vlad Mihailescu <vtmihailescu@gmail.com>