[Core] Update error message for Whisper + num-scheduler-steps > 1
Previously, using `--num-scheduler-steps` greater than 1 with an
encoder/decoder model like Whisper would result in the following
misleading error message:

```
ERROR 06-05 19:58:13 [engine.py:448] NotImplementedError: Speculative decoding is not currently supported with encoder/decoder models.
```

The reason is because `num_lookahead_slots` gets set based on the value
of `num-scheduler-steps` when spec decode is not in use. That logic is
here:

https://github.com/vllm-project/vllm/blob/8267f9916f9b4ceace3a8daa908d52824068862b/vllm/engine/arg_utils.py#L1154-L1160

```python
        # make sure num_lookahead_slots is set the higher value depending on
        # if we are using speculative decoding or multi-step
        num_lookahead_slots = max(self.num_lookahead_slots,
                                  self.num_scheduler_steps - 1)
        num_lookahead_slots = num_lookahead_slots \
            if speculative_config is None \
            else speculative_config.num_lookahead_slots
```

The error message for this case now looks like this:

```
ERROR 06-06 14:20:25 [engine.py:458] NotImplementedError: num-scheduler-steps > 1 is not currently supported with encoder/decoder models.
```

Signed-off-by: Russell Bryant <rbryant@redhat.com>