[Misc] Change tests/compile to use VLLM_V1 by default
Summary:
Previously, the tests were setting USE_VLLM_V1=0. Because torch.compile
is turned on by default for V1 (and off by default for V0), and because
we are moving to deprecate V0, this PR changes tests/compile to set
USE_VLLM_V1=1. Realistically speaking V1 and V0 both use compile in the
same way, so we end up testing the almost the same thing.

Test Plan:
Added an assertion that these tests are indeed being run with
USE_VLLM_V1=1.

Signed-off-by: rzou <zou3519@gmail.com>