[Bugfix] VLLM_V1 supports passing other compilation levels
Summary:
Before this PR, V1 always forced the compilation level to be 3
(PIECEWISE). In VLLM V0, users had an option to set the compilation
level to be something else. This PR changes V1 to respect the
compilation level set by the user.

The other compilation levels are mostly just used for debugging, but
they are very helpful for debugging. I've been trying to track down some
silent incorrectness and -O1 is using vLLM with stock torch.compile so
it helps in seeing if the problem is in stock torch.compile or vLLM.

Test Plan:
Added some new tests to test CompilationLevel 0 and 1. I'm not sure
exactly how to write a behavior test for CompilationLevel 2.

Signed-off-by: Richard Zou <zou3519@gmail.com>