[BugFix] Allow use_cudagraph to work with dyanmic VLLM_USE_V1
Summary:
If a test dynamically changes VLLM_USE_V1, the default value of
use_cudagraph doesn't get updated. This PR fixes that by setting a
default factory for use_cudagraph. I'm not sure if this is relevant in
real life but it is definitely relevant during testing (we dynamically
set VLLM_USE_V1 all the time in tests).

Test Plan:
- added new test (`pytest tests/compile/test_config.py -v -k "use_cudagraphs_dynamic"`)

Signed-off-by: rzou <zou3519@gmail.com>