[V1] Change return type on get_multimodal_embeddings()
The return type implied that `None` was a valid return value, but the
code does not treat it that way. `sanity_check_mm_encoder_outputs()` in
`vllm.v1.worker.utils` is called on the output and does not allow None.

While we're at it, also change the input parameter type to the sanity
check function. Previously, it was specified as `object`, but it should
be `MultiModalEmbeddings`.

Implementations of the method can still return an empty list if
necessary.

Signed-off-by: Russell Bryant <rbryant@redhat.com>