[Gaudi][Model] Qwen2.5-VL optimization for 112 aligned images (#1163)
This PR optimizes model Qwen2.5-VL for HPU with the assumption that
**images are 112 aligned**, meaning that height and width need to be
multiples of 112. If your image is not 112 aligned, we will
automatically trim it to comply with this requirement risking loosing
information. For example, an image of size `580 x 560` will be *trimmed*
to `560 x 560`.

If you don't want to resize the images we recommend using this custom
transformer version that automatically resizes the images:

``` 
pip install git+https://github.com/malkomes/transformers.git@ac372cd18f836c41f57cdce46094db00019d4280
``` 
To validate your setup, run 
```
pytest tests/models/multimodal/processing/test_qwen2_5_vl.py
```
to make sure that you have the correct version of transformers
installed.

Additionally, in this PR we:

- split the `visual` model encoder and the `language_model` graphs into
separate HPU Graphs. Notice that this separation of the visual encoder
is aligned with the optimizations done on V1, but here we do so because
of the dynamic operations of the visual model.
- we split the forward method on `VisionTransformer` into three
pre-attettion, forward, post-attention to isolate the dynamic
operations. To provide a modular code, we isolate these modification in
a custom class `Qwen2_5_VisionTransformerStaticShape`.
- We optimize full and window attention for HPU using
`AttentionLongSequence` and a special mask attention when appropriate.

-- 
Co-authored-by:  Gustavo Malkomes gustavo.malkomes@intel.com
Co-authored-by: Jimin Ha <jimin.ha@intel.com>
Co-authored-by: Sayantan Sarkar <sayantan.sarkar@intel.com>
Co-authored-by: Iman Gohari <s.m.iman.gohari@intel.com>

---------

Signed-off-by: Iman Gohari <s.m.iman.gohari@intel.com>
Signed-off-by: malkomes <74630871+malkomes@users.noreply.github.com>
Co-authored-by: Jimin Ha <jimin.ha@intel.com>
Co-authored-by: Gustavo Malkomes <gmalkomes@habana.ai>
Co-authored-by: Iman Gohari <s.m.iman.gohari@intel.com>
Co-authored-by: Michał Kuligowski <mkuligowski@habana.ai>
Co-authored-by: Sayantan Sarkar <sasarkar@habana.ai>