[doc] add project flag to gcloud TPU command
in case user doesn't have one set in gcloud config
which might be result in this error.

```
$ gcloud compute tpus tpu-vm ssh TPU_NAME --zone ZONE
ERROR: (gcloud.compute.tpus.tpu-vm.ssh) The required property [project] is not currently set.
It can be set on a per-command basis by re-running your command with the [--project] flag.

You may set it for your current workspace by running:

  $ gcloud config set project VALUE

or it can be set temporarily by the environment variable [CLOUDSDK_CORE_PROJECT]
```

Signed-off-by: David Xia <david@davidxia.com>