feat: add enable_force_include_usage option
Currently, when streaming the usage is
always null. This prevents the usage of
limits per user and is a bit unexpected as
without using streaming, the usage is always
returned.

This can be used, if vllm is between a
router, such as vllm-router or litellm.
and serves many users and is important to
detect abuse, divide costs etc.

Signed-off-by: Max Wittig <max.wittig@siemens.com>