[Core] More fixes to MultiModalEmbeddings type handling
This is a follow up to PR #19446.

In that PR, get_multimodal_embeddings() was changed to return
`MultiModalEmbeddings` instead of `Optional[MultiModalEmbeddings]`
because code in the model runner was requiring that the result was not
`None`.

Several models needed tweaks to account for this. Many were missed
because they were not tested in CI. This should fix the rest of the
common changes needed that weren't caught by CI.

Signed-off-by: Russell Bryant <rbryant@redhat.com>