[Feature] Constrain EPLB to main models
Otherwise, if EPLB is enabled, `FusedMoE` in MTP model
will also try to use expert mapping and record load metrics,
which is not wanted.

Later work should consider adding support for `EplbState`
to register multiple models, hence we can do EPLB on both
the main model and the MTP model.

Signed-off-by: Bowen Wang <abmfy@icloud.com>