test_attention compat with coming xformers change
Summary:
Update a test in vllm to not depend on internal details of the LowerTriangularMaskWithTensorBias object in xformers. This makes it work both with and without an expected upcoming change to xformers.

Testing: this is a small change to a vllm test which leaves it passing.

Context: the xformers library is expecting to change bias types back to not inherit tensors - basically https://github.com/facebookresearch/xformers/commit/96fd4808a4e976b0f30bff6d2f413df7d9d9a2d1 will be reverted. That inheritance is no longer needed for torch.compile, and in fact makes life harder for torch.export.

Test Plan:
See CI on D75689669

Rollback Plan:

Differential Revision: D77789850