Remove logic that uses more memory in prepare_attn_masks (#1597)
This has 2 changes
1. Inside prepare_attn_masks, there is extra code to address upstream
gemma3 attn_mask issue(see similar issue reported
https://github.com/huggingface/transformers/issues/39389). As this code
creates 2 extra tensors, it causes 128k run with OOM. We plan to remove
it for now as we don't see any accuracy drop with lmeval test. And plan
to fix it either when upstream fixes it or in different PR. Wo this
change, the text+image can't support >40k, with this change, can support
108k
3. when warmup multimodal vision bucket, there is no need to execute
warmup language model as prompt warmup is separate.

---------

Co-authored-by: Jimin Ha <jimin.ha@intel.com>