Add support for model signature verification on local models
Add support for model signature verification on local models. To do this,
replicate many of the command line options from the model signing library's
model_signing tool into vllm so that many use cases (different types of
signatures) are addressed.

For the local model signature verification check early on whether a
directory with the given name is available and if signature verification
is requested by the user by passing the --verification-method option.

Signature verification for models downloaded from the Huggingface hub
is not addressed in this patch.

To verify this patch you may want to try the following signed model:

git clone https://huggingface.co/ibm-granite/granite-4.0-tiny-preview

Verify and run it with the following command line:

vllm serve \
        --verification-method sigstore \
        --identity Granite.Preview@ibm.com \
        --identity-provider https://sigstore.verify.ibm.com/oauth2 \
        ~/models/granite-4.0-tiny-preview/

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>