Add support for model signature verification on local models
Add support for model signature verification on local models. To do this,
pass a security policy to vLLM that describes the models by their local
(absolute) path and their expected signers. An example of such a policy
looks like this:

  {
    "policy": {
      "signatures": {
        "signers": {
          "signer1": {
            "verification_method": "sigstore",
            "identity": "Granite.Preview@ibm.com",
            "identity_provider": "https://sigstore.verify.ibm.com/oauth2"
          }
        },
        "models": {
          "/tmp/granite-4.0-tiny-preview": {
            "signer": "signer1"
          }
        }
      }
    }
  }

This policy would only allow the one shown model to be loaded and it must
pass signature verification using the sigstore method with the given
identity and identity_provider. The above policy assumes the default model
signature file model.sig.

With the above policy in /tmp/policy.json, it is now possible to verify the
signature on a signed model, such as an IBM Granite model:

cd /tmp
git clone https://huggingface.co/ibm-granite/granite-4.0-tiny-preview

vllm serve \
	--security-policy /tmp/policy.json
        /tmp/granite-4.0-tiny-preview/

Note that signature verification of models downloaded by vLLM from
Huggingface is not addressed in this patch.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>