Drop flaky test_healthcheck_response_time
This test is flaky in CI.

https://buildkite.com/organizations/vllm/analytics/suites/ci-1/tests/c4322c0f-95c8-8ab8-a80f-01a25f114d5e?period=28days

This test is very timing sensitive and may fail depending on what
system it is run on, or what else is going on in the system under
test. If we ever try to run more tests in parallel, it will make this
worse.

One example failure has a message like this:

  [2025-08-04T10:45:00Z] >       assert load_response_time < 0.1
  [2025-08-04T10:45:00Z] E       assert 0.10481084599996393 < 0.1

The error here is that the response time took just over 0.1 seconds.

We could adjust the timing parameters, but I don't think adjusting to
where it won't be flaky will result in a test we get enough value out
of. I propose just dropping it.

Signed-off-by: Russell Bryant <rbryant@redhat.com>