vllm fix check on max vocab size (#22471)
Summary:

the tokenizer.vocab_size and model.vocab_size can be different. For QWen model, the tokenizer max token id is 151643 and the model config is `"vocab_size": 151936`. If we send an id between 151643 and 151936, it'll fail. Though in reality the tokenizer will just put ''. 

It's probably still valid to send the ids in between, because the model can legitimately produce such token id.

Test Plan:
Send 151860 and it's passing. Send 152860 and it complained about invalid token.

Rollback Plan:

Reviewed By: tensormeta

Differential Revision: D79840114