[Kernel][AMD] Avoid D2H copy and cumsum kernel
Current AITER attention for prefill incurs one D2H copy (which causes
synchonrization between CPU and GPU) and one cumsum kernel.

The D2H copy is used to get the actual KV tokens on the host side for
intermediate buffer allocation; this value should not change during the
inference, so we can pre-compute it during AttentionMetadata construction
time. And similarly, the cumsum kernel can be done in AttentionMedata
construction time

Signed-off-by: Xiaozhu <mxz297@gmail.com>