Fix cuda illegal mem access with Llama4 TP8 + rms_norm custom op
The rms_norm custom op cuda kernel supports non-contiguous striding on
the "num_tokens" dim, but it assumes that all the num_tokens dims (if
there are more than one num_tokens dim) are contiguous.

This causes a CUDA illegal mem access when running Llama4 with TP8 with
+rms_norm custom op.

To work around this issue, reshape the Q/K to 2D tensors before calling
the qk_norm. This makes the num_tokens dims collapse into one dim and it
will be obviously contiguous.

When rms_norm custom op is disabled, this reshape will be fused with the
RMSNorm's forward_native() anyway so there is no perf impact.

I will add additional checks in the custom rms_norm kernel in another
PR.

Signed-off-by: Po-Han Huang <pohanh@nvidia.com>