Improve multimodal hasher performance for re-used Image prompts
When processing batches of questions with multiple images each,
MultiModalHasher will re-compute the hash for every image in each
prompt. This is quite expensive and it blocks the ingress thread.

This change allows to apply an UUID tag to any Image. If detected,
Hasher will just use the UUID bytes rather than the actual image.

To avoid conflicts, only real UUID objects are supported, though
that could theoretically be opened to arbitrary inputs as well.
The ImageID ExifTag used here _theoretically_ could be used for
this purpose even then - it should be uniquely identifying the image.

This PR does not modify any Image loading code to use this, but
adds a test case to validate the functionality.

Signed-off-by: Staszek Pasko <staszek@gmail.com>