[Bugfix] Make unspecified --host bind to dual stack
Passing --host="" does not listen on IPv6 if it is available,
nor does it listen to all interfaces. This means that the
default vLLM server start in dual-stack and multi-home
environments will behave inconsistently from other classic
servers.

The standard code path for binding hosts in Python asyncio
create_server is that host=None binds to all addresses
on both IPv4 and IPv6, by retrieving all addresses that are
"passive" with socket.AI_PASSIVE. The create_server_socket
method only binds to a single address, and instead should
match the behavior of asyncio.loop.create_server (as it
previously did).

Change the create_server_socket method to create a list of
sockets and propagate that to callers.

Signed-off-by: Clayton Coleman <smarterclayton@gmail.com>