scheduler: don't leave request with partially allocated kv cache blocks
Previously we could have allocated blocks for cross-attention, but
then failed to allocate blocks for decoder self attention. We didn't
leak them, but we were holding them unnecessarily, when it's possible
they could be needed for something else. we will re-allocate from
scratch the next time we try scheduling the request.

Signed-off-by: Russell Bryant <rbryant@redhat.com>