Add missing max_qlen argument
Recently in https://github.com/ROCm/aiter.git new
argument was added to interface of pa_fwd_asm.

This is a breaking change since it changes the order
of arguments.

Apparently, the breaking change was introduced in
commit
[e44b95cf add pa qlen for mtp (#320)](https://github.com/ROCm/aiter/commit/e44b95cfaa80bf96c9790a7bea037801ca665573)
in aiter.

Without adding the missing argument calls to pa_fwd_asm
error with
`TypeError: max_qlen needs to be <class 'int'> but got <class 'torch.Tensor'>`

Alternative way to fix this would be to change aiter's interface.

Signed-off-by: tuukkjs <tuukka.sarvi@amd.com>