whisper: simplify encoder attention by using pytorch
We don't need the extra features of the attention backends available
through the more involved Attention abstraction. Replace encoder
attention with a new `TorchAttention` class which uses
scaled_dot_product_attention.

This also simplifies how the encoder can be run, as it no longer needs
to be run within a forward context with all of the attention metadata
available.

Signed-off-by: Russell Bryant <rbryant@redhat.com>