feat: Add deepseek v3.1 reasoning parser
DeepSeek v3.1 is conditionally thinking, so using the existing r1 parser
does not work correctly when a non-thinking request is submitted. This
adds a new parser that conditionally handles thinking when a request is
submitted with the `chat_template_kwargs` of `thinking` set to `true`.
(This follows the behavior of the chat template used by the model.)

To include the conditional behavior in the `extract_reasoning_content_streaming`
method, the `request` is added as an additional argument (this aligns
with prior changes to the `extract_reasoning_content` method to include
the `request` as an argument).

Signed-off-by: Andrew Senetar <asenetar@coreweave.com>