[#20711] Use QuantFp8 CustomOp-abstraction for MoE layers
Purpose: #19830 added QuantFp8, which uses the CustomOp abstraction to implement fp8 quantization in both CUDA and torch, allowing Inductor to achieve superior performance over the CUDA ops (which are unoptimized and also do not fuse by default). However, the class has to be instantiated during init, and MoE uses are currently in util free functions many levels deep. Those need to be mildly rearchitected to take advantage of the new abstraction.

The main changes include:
- Add a MoEInputQuantizer class with init/forward (and callable) interface for FP8/INT8/NVFP4/MXFP4 activation quantization.
	- Make the class abstract out details on decision of the quant fp8 op to be used

Signed-off-by: Rohan Jagtap <rohanj30@icloud.com>