[#20711] Use QuantFp8 CustomOp-abstraction for MoE layers
Purpose: #19830 added QuantFp8, which uses the CustomOp abstraction to implement fp8 quantization in both CUDA and torch, allowing Inductor to achieve superior performance over the CUDA ops (which are unoptimized and also do not fuse by default). However, the class has to be instantiated during init, and MoE uses are currently in util free functions many levels deep. Those need to be mildly rearchitected to take advantage of the new abstraction.

The changes in this commit include:
- Refactor docstrings to use backticks for code
- refactor formatting for certain methods to match other calls

Signed-off-by: Rohan Jagtap <rohanj30@icloud.com>