[ux] Switch a warning to debug about a pytorch fallback
This changes the following log message from WARNING to DEBUG.

  WARNING 08-27 07:58:39 [topk_topp_sampler.py:100]
    FlashInfer 0.2.3+ does not support per-request generators.
    Falling back to PyTorch-native implementation.

Since this is the code working as intended and there is nothing a user
can do with the information, it seems better to keep it as DEBUG level
information.

I made this change after seeing a question from someone about how to
get the message to go away.

Signed-off-by: Russell Bryant <rbryant@redhat.com>