Fix test failure by keeping lora_extra_vocab_size default at 256
The previous commit changed the default value from 256 to 0, which
caused test_v1_llm_by_default to fail. This commit keeps the original
default value of 256 while still adding the deprecation warning to
notify users that this feature will be removed in v0.12.0.

The deprecation warning will always be shown when LoRAConfig is created,
alerting users to the upcoming removal without breaking existing code.

Signed-off-by: Jinheng Li <ahengljh@gmail.com>