[kernel] Add stride checks for rms_norm kernels
This is a follow-up PR from #22701

Currently, the rms_norm kernels (including the ones with fused
ResidualAdd and/or FP8-Quant) supported partially strided tensors, with
the following restrictions:

- The stride of the last dimension (hidden_size) must be 1.
- The num_tokens dimensions (if more than 1) must be contiguous.

Previously, we did not check these in the custom op implementation. That
caused silent corruptions and cuda illegal memory access if users passed
in tensors not supported by the kernels.

This PR adds such checks to ensure that when this situation happens, it
fails loudly with corresponding error messages.

The unit test (test_layernorm.py) is also updated to add such "negative"
(i.e. unsupported) test cases and make sure the custom ops raise an
error when users pass in unsupported tensors.

This PR also fuses the rms_norm tests with rms_norm+quant tests since
the two are very similar and have duplicated code.

Signed-off-by: Po-Han Huang <pohanh@nvidia.com>